chg: refactor WprRecord, WprReplay to WprAbstract
